+++ src/core.c
@@ -2,11 +2,15 @@
 
 #include "sakura.h"
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
 // internal functions ----------------------------------------------------------
 
 static SEXP nano_eval_res;
 
-static void nano_eval_safe (void *call) {
+static void nano_eval_safe(void *call) {
   nano_eval_res = Rf_eval((SEXP) call, R_GlobalEnv);
 }
 
@@ -34,9 +38,157 @@ static void nano_write_bytes(R_outpstream_t stream, void *src, int len) {
 static void nano_read_bytes(R_inpstream_t stream, void *dst, int len) {
 
   nano_buf *buf = (nano_buf *) stream->data;
-  if (buf->cur + len > buf->len) Rf_error("unserialization error");
+  if (buf->cur + (size_t) len > buf->len) Rf_error("unserialization error");
   memcpy(dst, buf->buf + buf->cur, len);
-  buf->cur += len;
+  buf->cur += (size_t) len;
+
+}
+
+typedef struct sakura_gz_cleanup_s {
+  const char *path;
+  sakura_gz_file *file;
+} sakura_gz_cleanup;
+
+typedef struct sakura_save_rds_data_s {
+  sakura_gz_cleanup cleanup;
+  SEXP object;
+  SEXP config;
+} sakura_save_rds_data;
+
+typedef struct sakura_read_rds_data_s {
+  sakura_gz_cleanup cleanup;
+} sakura_read_rds_data;
+
+static void sakura_serialize_stream(
+  R_outpstream_t output_stream,
+  R_pstream_data_t data,
+  void (*outbytes)(R_outpstream_t, void *, int),
+  SEXP object,
+  SEXP config
+);
+
+static SEXP sakura_unserialize_stream(
+  R_inpstream_t input_stream,
+  R_pstream_data_t data,
+  void (*inbytes)(R_inpstream_t, void *, int)
+);
+
+static void sakura_write_gzip_bytes(R_outpstream_t stream, void *src, int len) {
+
+  sakura_gz_file *file = (sakura_gz_file *) stream->data;
+  unsigned char *cursor = (unsigned char *) src;
+
+  while (len > 0) {
+    int written = gzwrite(file->file, cursor, (unsigned int) len);
+    if (written <= 0) {
+      int errnum = Z_OK;
+      const char *msg = gzerror(file->file, &errnum);
+      Rf_error("gzip write error: %s", msg);
+    }
+    cursor += written;
+    len -= written;
+  }
+
+}
+
+static void sakura_read_gzip_bytes(R_inpstream_t stream, void *dst, int len) {
+
+  sakura_gz_file *file = (sakura_gz_file *) stream->data;
+  unsigned char *cursor = (unsigned char *) dst;
+
+  while (len > 0) {
+    int read = gzread(file->file, cursor, (unsigned int) len);
+    if (read <= 0) {
+      int errnum = Z_OK;
+      const char *msg = gzerror(file->file, &errnum);
+      if (errnum == Z_OK && gzeof(file->file)) {
+        Rf_error("unexpected end of gzip stream");
+      }
+      Rf_error("gzip read error: %s", msg);
+    }
+    cursor += read;
+    len -= read;
+  }
+
+}
+
+static void sakura_gz_cleanup_close(void *data, Rboolean jump) {
+
+  sakura_gz_cleanup *cleanup = (sakura_gz_cleanup *) data;
+
+  if (cleanup->file->file != NULL) {
+    int status = gzclose(cleanup->file->file);
+    cleanup->file->file = NULL;
+
+    if (!jump && status != Z_OK) {
+      Rf_error("failed to close gzip file '%s'", cleanup->path);
+    }
+  }
+
+}
+
+static void nano_write_string(R_outpstream_t stream, SEXP string) {
+
+  int length = LENGTH(string);
+  const char *value = CHAR(string);
+  void (*OutBytes)(R_outpstream_t, void *, int) = stream->OutBytes;
+
+  OutBytes(stream, &length, sizeof(int));
+  if (length) {
+    // OutBytes uses a non-const pointer even though the bytes are not mutated.
+    OutBytes(stream, (void *) value, length);
+  }
+
+}
+
+static SEXP nano_read_string(R_inpstream_t stream) {
+
+  int length;
+  char *value;
+  void (*InBytes)(R_inpstream_t, void *, int) = stream->InBytes;
+
+  InBytes(stream, &length, sizeof(int));
+  if (length < 0) {
+    Rf_error("unserialization error");
+  }
+
+  value = (char *) R_alloc((size_t) length + 1, sizeof(char));
+  if (length) {
+    InBytes(stream, value, length);
+  }
+  value[length] = '\0';
+
+  return Rf_mkString(value);
+
+}
+
+static SEXP sakura_save_rds_impl(void *data) {
+
+  sakura_save_rds_data *save_data = (sakura_save_rds_data *) data;
+  struct R_outpstream_st output_stream;
+
+  sakura_serialize_stream(
+    &output_stream,
+    (R_pstream_data_t) save_data->cleanup.file,
+    sakura_write_gzip_bytes,
+    save_data->object,
+    save_data->config
+  );
+
+  return R_NilValue;
+
+}
+
+static SEXP sakura_read_rds_impl(void *data) {
+
+  sakura_read_rds_data *read_data = (sakura_read_rds_data *) data;
+  struct R_inpstream_st input_stream;
+
+  return sakura_unserialize_stream(
+    &input_stream,
+    (R_pstream_data_t) read_data->cleanup.file,
+    sakura_read_gzip_bytes
+  );
 
 }
 
@@ -45,7 +197,8 @@ static SEXP nano_serialize_hook(SEXP x, SEXP bundle_xptr) {
   sakura_serial_bundle *bundle = (sakura_serial_bundle *) R_ExternalPtrAddr(bundle_xptr);
   R_outpstream_t stream = bundle->stream;
   SEXP klass = bundle->klass;
-  SEXP hook_func = bundle->hook_func;
+  SEXP package = bundle->package;
+  SEXP sfunc = bundle->sfunc;
   int len = (int) XLENGTH(klass), match = 0, i;
   void (*OutBytes)(R_outpstream_t, void *, int) = stream->OutBytes;
 
@@ -60,16 +213,15 @@ static SEXP nano_serialize_hook(SEXP x, SEXP bundle_xptr) {
     return R_NilValue;
 
   SEXP call;
-  PROTECT(call = Rf_lcons(VECTOR_ELT(hook_func, i), Rf_cons(x, R_NilValue)));
+  PROTECT(call = Rf_lcons(VECTOR_ELT(sfunc, i), Rf_cons(x, R_NilValue)));
   if (!R_ToplevelExec(nano_eval_safe, call) || TYPEOF(nano_eval_res) != RAWSXP) {
-    // something went wrong
     UNPROTECT(1);
     return R_NilValue;
   }
   UNPROTECT(1);
 
   // write out PERSISTSXP manually
-  uint64_t size = XLENGTH(nano_eval_res);
+  uint64_t size = (uint64_t) XLENGTH(nano_eval_res);
   char size_string[21];
   snprintf(size_string, sizeof(size_string), "%020" PRIu64, size);
 
@@ -87,7 +239,7 @@ static SEXP nano_serialize_hook(SEXP x, SEXP bundle_xptr) {
   OutBytes(stream, &int_1, sizeof(int));          // 12
   OutBytes(stream, &int_charsxp, sizeof(int));    // 16
   OutBytes(stream, &int_20, sizeof(int));         // 20
-  OutBytes(stream, &size_string[0], 20);      // 40
+  OutBytes(stream, &size_string[0], 20);          // 40
 
   // write out binary serialization blob
   unsigned char *dest = RAW(nano_eval_res);
@@ -97,8 +249,9 @@ static SEXP nano_serialize_hook(SEXP x, SEXP bundle_xptr) {
     size -= SAKURA_CHUNK_SIZE;
   }
   OutBytes(stream, dest, (int) size);
-  // write out index if hook_func is a vector
-  OutBytes(stream, &i, sizeof(int));      // 4
+
+  nano_write_string(stream, STRING_ELT(klass, i));
+  nano_write_string(stream, STRING_ELT(package, i));
 
   return R_BlankScalarString; // this will write the PERSISTSXP again
 
@@ -108,14 +261,13 @@ static SEXP nano_unserialize_hook(SEXP x, SEXP bundle_xptr) {
 
   sakura_unserial_bundle *bundle = (sakura_unserial_bundle *) R_ExternalPtrAddr(bundle_xptr);
   R_inpstream_t stream = bundle->stream;
-  SEXP hook_func = bundle->hook_func;
   void (*InBytes)(R_inpstream_t, void *, int) = stream->InBytes;
 
   const char *size_string = CHAR(STRING_ELT(x, 0));
-  uint64_t size = strtoul(size_string, NULL, 10);
+  uint64_t size = (uint64_t) strtoull(size_string, NULL, 10);
 
-  SEXP raw, call, out;
-  PROTECT(raw = Rf_allocVector(RAWSXP, size));
+  SEXP raw, class_name, package_name, name, ns, call, out;
+  PROTECT(raw = Rf_allocVector(RAWSXP, (R_xlen_t) size));
   unsigned char *dest = RAW(raw);
   while (size > SAKURA_CHUNK_SIZE) {
     InBytes(stream, dest, SAKURA_CHUNK_SIZE);
@@ -124,30 +276,125 @@ static SEXP nano_unserialize_hook(SEXP x, SEXP bundle_xptr) {
   }
   InBytes(stream, dest, (int) size);
 
-  int i;
-  InBytes(stream, &i, 4);
+  PROTECT(class_name = nano_read_string(stream));
+  PROTECT(package_name = nano_read_string(stream));
 
-  // read in 20 additional bytes and discard them
-  char buf[20];
-  InBytes(stream, buf, 20);
+  // read in the automatic trailing PERSISTSXP bytes and discard them
+  char discard[SAKURA_PERSIST_DISCARD];
+  InBytes(stream, discard, SAKURA_PERSIST_DISCARD);
 
-  PROTECT(call = Rf_lcons(VECTOR_ELT(hook_func, i), Rf_cons(raw, R_NilValue)));
-  out = Rf_eval(call, R_GlobalEnv);
+  PROTECT(name = Rf_mkString("sakura"));
+  PROTECT(ns = R_FindNamespace(name));
+  PROTECT(call = Rf_lcons(
+    Rf_install(".sakura_dispatch_unserialize"),
+    Rf_cons(
+      raw,
+      Rf_cons(class_name, Rf_cons(package_name, R_NilValue))
+    )
+  ));
+  out = Rf_eval(call, ns);
 
-  UNPROTECT(2);
+  UNPROTECT(6);
   return out;
+
+}
+
+static void sakura_serialize_stream(
+  R_outpstream_t output_stream,
+  R_pstream_data_t data,
+  void (*outbytes)(R_outpstream_t, void *, int),
+  SEXP object,
+  SEXP config
+) {
+
+  if (config == R_NilValue) {
+
+    R_InitOutPStream(
+      output_stream,
+      data,
+      R_pstream_binary_format,
+      SAKURA_SERIAL_VER,
+      NULL,
+      outbytes,
+      NULL,
+      R_NilValue
+    );
+
+    R_Serialize(object, output_stream);
+
+  } else {
+
+    sakura_serial_bundle bundle;
+    R_SetExternalPtrAddr(sakura_bundle, &bundle);
+
+    sakura_serialize_init(
+      sakura_bundle,
+      output_stream,
+      data,
+      VECTOR_ELT(config, 0),
+      VECTOR_ELT(config, 1),
+      VECTOR_ELT(config, 2),
+      outbytes
+    );
+
+    R_Serialize(object, output_stream);
+    R_SetExternalPtrAddr(sakura_bundle, NULL);
+
+  }
+
+}
+
+static SEXP sakura_unserialize_stream(
+  R_inpstream_t input_stream,
+  R_pstream_data_t data,
+  void (*inbytes)(R_inpstream_t, void *, int)
+) {
+
+  sakura_unserial_bundle bundle;
+  SEXP out;
+  R_SetExternalPtrAddr(sakura_bundle, &bundle);
+
+  sakura_unserialize_init(
+    sakura_bundle,
+    input_stream,
+    data,
+    inbytes
+  );
+
+  out = R_Unserialize(input_stream);
+  R_SetExternalPtrAddr(sakura_bundle, NULL);
+
+  return out;
+
+}
+
+static const char *sakura_file_path(SEXP file) {
+
+  if (TYPEOF(file) != STRSXP || XLENGTH(file) != 1 || STRING_ELT(file, 0) == NA_STRING) {
+    Rf_error("`file` must be a character string");
+  }
+
+  return R_ExpandFileName(CHAR(STRING_ELT(file, 0)));
 
 }
 
 // C API functions -------------------------------------------------------------
 
-void sakura_serialize_init(SEXP bundle_xptr, R_outpstream_t stream, R_pstream_data_t data,
-                           SEXP klass, SEXP hook_func, void (*outbytes)(R_outpstream_t, void *, int)) {
+void sakura_serialize_init(
+  SEXP bundle_xptr,
+  R_outpstream_t stream,
+  R_pstream_data_t data,
+  SEXP klass,
+  SEXP package,
+  SEXP sfunc,
+  void (*outbytes)(R_outpstream_t, void *, int)
+) {
 
   sakura_serial_bundle *bundle = (sakura_serial_bundle *) R_ExternalPtrAddr(bundle_xptr);
 
   bundle->klass = klass;
-  bundle->hook_func = hook_func;
+  bundle->package = package;
+  bundle->sfunc = sfunc;
   bundle->stream = stream;
   R_InitOutPStream(
     stream,
@@ -162,12 +409,15 @@ void sakura_serialize_init(SEXP bundle_xptr, R_outpstream_t stream, R_pstream_da
 
 }
 
-void sakura_unserialize_init(SEXP bundle_xptr, R_inpstream_t stream, R_pstream_data_t data,
-                             SEXP hook_func, void (*inbytes)(R_inpstream_t, void *, int)) {
+void sakura_unserialize_init(
+  SEXP bundle_xptr,
+  R_inpstream_t stream,
+  R_pstream_data_t data,
+  void (*inbytes)(R_inpstream_t, void *, int)
+) {
 
   sakura_unserial_bundle *bundle = (sakura_unserial_bundle *) R_ExternalPtrAddr(bundle_xptr);
 
-  bundle->hook_func = hook_func;
   bundle->stream = stream;
   R_InitInPStream(
     stream,
@@ -181,7 +431,7 @@ void sakura_unserialize_init(SEXP bundle_xptr, R_inpstream_t stream, R_pstream_d
 
 }
 
-void sakura_serialize(nano_buf *buf, SEXP object, SEXP hook) {
+void sakura_serialize(nano_buf *buf, SEXP object, SEXP config) {
 
   buf->buf = R_Calloc(SAKURA_INIT_BUFSIZE, unsigned char);
   buf->len = SAKURA_INIT_BUFSIZE;
@@ -189,43 +439,17 @@ void sakura_serialize(nano_buf *buf, SEXP object, SEXP hook) {
 
   struct R_outpstream_st output_stream;
 
-  if (hook == R_NilValue) {
-
-    R_InitOutPStream(
-      &output_stream,
-      (R_pstream_data_t) buf,
-      R_pstream_binary_format,
-      SAKURA_SERIAL_VER,
-      NULL,
-      nano_write_bytes,
-      NULL,
-      R_NilValue
-    );
-
-    R_Serialize(object, &output_stream);
-
-  } else {
-
-    sakura_serial_bundle b;
-    R_SetExternalPtrAddr(sakura_bundle, &b);
-
-    sakura_serialize_init(
-      sakura_bundle,
-      &output_stream,
-      (R_pstream_data_t) buf,
-      VECTOR_ELT(hook, 0),
-      VECTOR_ELT(hook, 1),
-      nano_write_bytes
-    );
-
-    R_Serialize(object, &output_stream);
-    R_SetExternalPtrAddr(sakura_bundle, NULL);
-
-  }
+  sakura_serialize_stream(
+    &output_stream,
+    (R_pstream_data_t) buf,
+    nano_write_bytes,
+    object,
+    config
+  );
 
 }
 
-SEXP sakura_unserialize(unsigned char *buf, size_t sz, SEXP hook) {
+SEXP sakura_unserialize(unsigned char *buf, size_t sz) {
 
   nano_buf nbuf;
   nbuf.buf = buf;
@@ -233,50 +457,21 @@ SEXP sakura_unserialize(unsigned char *buf, size_t sz, SEXP hook) {
   nbuf.cur = 0;
 
   struct R_inpstream_st input_stream;
-  SEXP out;
 
-  if (hook == R_NilValue) {
-
-    R_InitInPStream(
-      &input_stream,
-      (R_pstream_data_t) &nbuf,
-      R_pstream_any_format,
-      NULL,
-      nano_read_bytes,
-      NULL,
-      R_NilValue
-    );
-
-    out = R_Unserialize(&input_stream);
-
-  } else {
-
-    sakura_unserial_bundle b;
-    R_SetExternalPtrAddr(sakura_bundle, &b);
-
-    sakura_unserialize_init(
-      sakura_bundle,
-      &input_stream,
-      (R_pstream_data_t) &nbuf,
-      VECTOR_ELT(hook, 2),
-      nano_read_bytes
-    );
-
-    out = R_Unserialize(&input_stream);
-    R_SetExternalPtrAddr(sakura_bundle, NULL);
-
-  }
-
-  return out;
+  return sakura_unserialize_stream(
+    &input_stream,
+    (R_pstream_data_t) &nbuf,
+    nano_read_bytes
+  );
 
 }
 
 // R API functions -------------------------------------------------------------
 
-SEXP sakura_r_serialize(SEXP object, SEXP hook) {
+SEXP sakura_r_serialize(SEXP object, SEXP config) {
 
   nano_buf buf;
-  sakura_serialize(&buf, object, hook);
+  sakura_serialize(&buf, object, config);
 
   SEXP out = Rf_allocVector(RAWSXP, buf.cur);
   memcpy(RAW(out), buf.buf, buf.cur);
@@ -286,8 +481,60 @@ SEXP sakura_r_serialize(SEXP object, SEXP hook) {
 
 }
 
-SEXP sakura_r_unserialize(SEXP object, SEXP hook) {
+SEXP sakura_r_unserialize(SEXP object) {
 
-  return sakura_unserialize(RAW(object), XLENGTH(object), hook);
+  return sakura_unserialize(RAW(object), XLENGTH(object));
+
+}
+
+SEXP sakura_r_save_rds(SEXP object, SEXP file, SEXP config) {
+
+  const char *path = sakura_file_path(file);
+  sakura_gz_file gz_file;
+  sakura_save_rds_data save_data;
+
+  gz_file.file = gzopen(path, "wb");
+  if (gz_file.file == NULL) {
+    Rf_error("failed to open file '%s' for writing", path);
+  }
+
+  save_data.cleanup.path = path;
+  save_data.cleanup.file = &gz_file;
+  save_data.object = object;
+  save_data.config = config;
+
+  R_UnwindProtect(
+    sakura_save_rds_impl,
+    &save_data,
+    sakura_gz_cleanup_close,
+    &save_data.cleanup,
+    NULL
+  );
+
+  return R_NilValue;
+
+}
+
+SEXP sakura_r_read_rds(SEXP file) {
+
+  const char *path = sakura_file_path(file);
+  sakura_gz_file gz_file;
+  sakura_read_rds_data read_data;
+
+  gz_file.file = gzopen(path, "rb");
+  if (gz_file.file == NULL) {
+    Rf_error("failed to open file '%s' for reading", path);
+  }
+
+  read_data.cleanup.path = path;
+  read_data.cleanup.file = &gz_file;
+
+  return R_UnwindProtect(
+    sakura_read_rds_impl,
+    &read_data,
+    sakura_gz_cleanup_close,
+    &read_data.cleanup,
+    NULL
+  );
 
 }